add atomic-internal.h
